@@ -1,10 +1,7 @@
 module github.com/Predixxion/gosoap
 
-require (
-	github.com/google/go-cmp v0.5.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/net v0.0.0-20210726213435-c6fcb2dbf985
-	gotest.tools v2.2.0+incompatible
-)
+go 1.23.2
 
-go 1.22
+require golang.org/x/net v0.30.0
+
+require golang.org/x/text v0.19.0 // indirect
